+++ src/gwt/src/org/rstudio/studio/client/workbench/views/source/editors/text/TextEditingTargetWidget.java
@@ -280,6 +280,7 @@ public class TextEditingTargetWidget
       toolbar.addLeftSeparator();
       toolbar.addLeftWidget(commands_.synctexSearch().createToolbarButton());
 
+      toolbar.addRightWidget(insertChunkButton_ = commands_.insertChunk().createToolbarButton());
       toolbar.addRightWidget(runButton_ = commands_.executeCode().createToolbarButton(false));
       toolbar.addRightSeparator();
       toolbar.addRightWidget(runLastButton_ = commands_.executeLastCode().createToolbarButton(false));
@@ -326,10 +327,6 @@ public class TextEditingTargetWidget
       //toolbar.addRightSeparator();
      
       ToolbarPopupMenu chunksMenu = new ToolbarPopupMenu();
-      chunksMenu.addItem(commands_.insertChunk().createMenuItem(false));
-      chunksMenu.addSeparator();
-      chunksMenu.addItem(commands_.jumpTo().createMenuItem(false));
-      chunksMenu.addSeparator();
       chunksMenu.addItem(commands_.executeSetupChunk().createMenuItem(false));
       chunksMenu.addItem(commands_.executePreviousChunks().createMenuItem(false));
       chunksMenu.addItem(commands_.executeCurrentChunk().createMenuItem(false));
@@ -337,8 +334,8 @@ public class TextEditingTargetWidget
       chunksMenu.addSeparator();
       chunksMenu.addItem(commands_.executeAllCode().createMenuItem(false));
       chunksButton_ = new ToolbarButton(
-                       "Chunks",  
-                       StandardIcons.INSTANCE.chunk_menu(), 
+                       "Run",  
+                       commands_.executeCode().getImageResource(),
                        chunksMenu, 
                        true);
       toolbar.addRightWidget(chunksButton_);
@@ -497,8 +494,11 @@ public class TextEditingTargetWidget
       boolean canPreviewFromR = fileType.canPreviewFromR();
       
       // don't show the run buttons for cpp files, or R files in Shiny
-      runButton_.setVisible(canExecuteCode && !isCpp && !isShinyFile());
+      runButton_.setVisible(canExecuteCode && !canExecuteChunks && !isCpp && !isShinyFile());
       runLastButton_.setVisible(runButton_.isVisible() && !canExecuteChunks);
+      
+      // show insert chunk button
+      insertChunkButton_.setVisible(canExecuteChunks);
       
       sourceOnSave_.setVisible(canSourceOnSave);
       srcOnSaveLabel_.setVisible(canSourceOnSave);
@@ -606,7 +606,7 @@ public class TextEditingTargetWidget
       else
          srcOnSaveLabel_.setText(width < 450 ? "Source" : "Source on Save");
       sourceButton_.setText(width < 400 ? "" : sourceCommandText_);
-      chunksButton_.setText(width < 400 ? "" : "Chunks");
+      chunksButton_.setText(width < 400 ? "" : "Run");
    }
    
    
@@ -1004,6 +1004,7 @@ public class TextEditingTargetWidget
    private ToolbarButton compilePdfButton_;
    private ToolbarButton previewHTMLButton_;
    private ToolbarButton knitDocumentButton_;
+   private ToolbarButton insertChunkButton_;
    private ToolbarButton runButton_;
    private ToolbarButton runLastButton_;
    private ToolbarButton sourceButton_;